requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy
opencv-python
Pillow
albumentations
segmentation-models-pytorch
timm
six
PyQt5
mlflow
matplotlib